Update keys if expired
If a key has expired it will still be shown in `apt-key list` but
instead of "expires" it will show "expired", so adding that to grep.

Sample outputs:
```
pub   4096R/ACCC4CF8 2011-10-13 [expired: 2016-02-24]

vs

pub   4096R/ACCC4CF8 2011-10-13 [expires: 2019-07-02]
```

Worth noting also that the key size may change, if so, some of the
grep commands would need to be changed too.